// src/api/models/CatalogTextNew.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SITE API
 * API сайта ПИК Брокер
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime'
import { Validation, ValidationFromJSON, ValidationFromJSONTyped, ValidationToJSON } from './'

/**
 *
 * @export
 * @interface CatalogTextNew
 */
export interface CatalogTextNew {
  /**
   *
   * @type {string}
   * @memberof CatalogTextNew
   */
  title: string
  /**
   *
   * @type {string}
   * @memberof CatalogTextNew
   */
  placeholder: string | null
  /**
   *
   * @type {string}
   * @memberof CatalogTextNew
   */
  type: CatalogTextNewTypeEnum
  /**
   *
   * @type {string}
   * @memberof CatalogTextNew
   */
  initial: string | null
  /**
   *
   * @type {Validation}
   * @memberof CatalogTextNew
   */
  validation: Validation
}

/**
 * @export
 * @enum {string}
 */
export enum CatalogTextNewTypeEnum {
  Text = 'text',
}

export function CatalogTextNewFromJSON(json: any): CatalogTextNew {
  return CatalogTextNewFromJSONTyped(json, false)
}

export function CatalogTextNewFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CatalogTextNew {
  if (json === undefined || json === null) {
    return json
  }
  return {
    title: json['title'],
    placeholder: json['placeholder'],
    type: json['type'],
    initial: json['initial'],
    validation: ValidationFromJSON(json['validation']),
  }
}

export function CatalogTextNewToJSON(value?: CatalogTextNew | null): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {
    title: value.title,
    placeholder: value.placeholder,
    type: value.type,
    initial: value.initial,
    validation: ValidationToJSON(value.validation),
  }
}